Fix gcc warnings on mingw and IRIX 6.5.
* gettext-tools/src/msgexec.c (process_message): Don't assume that fprintf
supports %zu.
* gettext-tools/src/msgfilter.c (process_message): Likewise.

+++ gettext-tools/src/msgexec.c
@@ -458,7 +458,7 @@ process_message (const message_ty *mp)
 
       if (mp->msgid_plural != NULL)
         {
-          char *plural_form_string = xasprintf ("%zu", k);
+          char *plural_form_string = xasprintf ("%lu", (unsigned long) k);
 
           xsetenv ("MSGEXEC_PLURAL_FORM", plural_form_string, 1);
           free (plural_form_string);
