+++ gettext-tools/src/msgfilter.c
@@ -739,7 +739,7 @@ process_message (message_ty *mp)
 
       if (mp->msgid_plural != NULL)
         {
-          char *plural_form_string = xasprintf ("%zu", k);
+          char *plural_form_string = xasprintf ("%lu", (unsigned long) k);
 
           xsetenv ("MSGFILTER_PLURAL_FORM", plural_form_string, 1);
           free (plural_form_string);
